backend/requirements.txt
# LexiFlow Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
pydub==0.25.1
aiofiles==23.2.1
